requirements.txt
# DESK — Python dependencies (pinned)
# Add per increment. Never add without flagging to Dhoni.

# Increment 1 — BigQuery schema setup
google-cloud-bigquery==3.25.0
pyarrow==16.1.0
db-dtypes==1.3.0
google-auth==2.35.0

# Increment 2-5 — Ingestion pipeline
requests==2.32.3
jsonschema==4.23.0

# Increment 6-8 — dbt transformations
dbt-core==1.8.7
dbt-bigquery==1.8.2
